[Step2] refactor : 코드리뷰 반영 - validation이 필요한 객체에서 검증
